+++ libraries/base/Data/Type/Bool.hs
@@ -1,6 +1,5 @@
-{-# LANGUAGE Safe #-}
-{-# LANGUAGE TypeFamilies, TypeOperators, DataKinds, NoImplicitPrelude,
-             PolyKinds #-}
+{-# LANGUAGE TypeFamilyDependencies, Safe, PolyKinds #-}
+{-# LANGUAGE TypeFamilies, TypeOperators, DataKinds, NoImplicitPrelude #-}
 
 -----------------------------------------------------------------------------
 -- |
@@ -49,7 +48,9 @@ type family a || b where
   a      || a      = a
 infixr 2 ||
 
--- | Type-level "not"
-type family Not a where
+-- | Type-level "not". An injective type family since @4.10.0.0@.
+--
+-- @since 4.7.0.0
+type family Not a = res | res -> a where
   Not 'False = 'True
   Not 'True  = 'False